agents/requirements.txt
agent-framework-azure-ai
agent-framework-devui
python-dotenv
azure-identity
